requirements.txt
requests==2.31.0
urllib3<2.0.0  # Use urllib3 v1.x for compatibility with LibreSSL on macOS
beautifulsoup4==4.12.2
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.1
# lxml removed - using built-in html.parser instead for better Windows compatibility
# lxml>=5.0.0
Pillow>=10.3.0
PyYAML==6.0.1
pytz==2023.3
selenium==4.36.0

